[refactor] 자주 사용되는 버튼 이름 변경
